refactoring/splitting the json fix functions into general module and llm module which need AI's assistance.
