+++ autogpt/json_utils/json_fix_llm.py
@@ -1,15 +1,15 @@
-"""This module contains the function to fix JSON strings"""
+"""This module contains functions to fix JSON strings generated by LLM models, such as ChatGPT, using the assistance
+of the ChatGPT API or LLM models."""
 from __future__ import annotations
 
 import contextlib
 import json
-import re
-from typing import Optional, Dict, Any
+from typing import Dict, Any
 
 from colorama import Fore
 from regex import regex
 
-from autogpt.json_utils.utilities import extract_char_position
+from autogpt.json_utils.json_fix_general import correct_json
 from autogpt.llm_utils import call_ai_function
 from autogpt.logs import logger
 from autogpt.speech import say_text
@@ -82,83 +82,6 @@ def auto_fix_json(json_string: str, schema: str) -> str:
         return "failed"
 
 
-def fix_invalid_escape(json_to_load: str, error_message: str) -> str:
-    """Fix invalid escape sequences in JSON strings.
-
-    Args:
-        json_to_load (str): The JSON string.
-        error_message (str): The error message from the JSONDecodeError
-          exception.
-
-    Returns:
-        str: The JSON string with invalid escape sequences fixed.
-    """
-    while error_message.startswith("Invalid \\escape"):
-        bad_escape_location = extract_char_position(error_message)
-        json_to_load = (
-                json_to_load[:bad_escape_location] + json_to_load[bad_escape_location + 1:]
-        )
-        try:
-            json.loads(json_to_load)
-            return json_to_load
-        except json.JSONDecodeError as e:
-            if CFG.debug_mode:
-                print("json loads error - fix invalid escape", e)
-            error_message = str(e)
-    return json_to_load
-
-
-def balance_braces(json_string: str) -> Optional[str]:
-    """
-    Balance the braces in a JSON string.
-
-    Args:
-        json_string (str): The JSON string.
-
-    Returns:
-        str: The JSON string with braces balanced.
-    """
-
-    open_braces_count = json_string.count("{")
-    close_braces_count = json_string.count("}")
-
-    while open_braces_count > close_braces_count:
-        json_string += "}"
-        close_braces_count += 1
-
-    while close_braces_count > open_braces_count:
-        json_string = json_string.rstrip("}")
-        close_braces_count -= 1
-
-    with contextlib.suppress(json.JSONDecodeError):
-        json.loads(json_string)
-        return json_string
-
-
-def add_quotes_to_property_names(json_string: str) -> str:
-    """
-    Add quotes to property names in a JSON string.
-
-    Args:
-        json_string (str): The JSON string.
-
-    Returns:
-        str: The JSON string with quotes added to property names.
-    """
-
-    def replace_func(match: re.Match) -> str:
-        return f'"{match[1]}":'
-
-    property_name_pattern = re.compile(r"(\w+):")
-    corrected_json_string = property_name_pattern.sub(replace_func, json_string)
-
-    try:
-        json.loads(corrected_json_string)
-        return corrected_json_string
-    except json.JSONDecodeError as e:
-        raise e
-
-
 def fix_json_using_multiple_techniques(assistant_reply: str) -> Dict[Any, Any]:
     """Fix the given JSON string to make it parseable and fully compliant with two techniques.
 
@@ -184,40 +107,6 @@ def fix_json_using_multiple_techniques(assistant_reply: str) -> Dict[Any, Any]:
         say_text("I have received an invalid JSON response from the OpenAI API.")
 
     return {}
-
-
-def correct_json(json_to_load: str) -> str:
-    """
-    Correct common JSON errors.
-    Args:
-        json_to_load (str): The JSON string.
-    """
-
-    try:
-        if CFG.debug_mode:
-            print("json", json_to_load)
-        json.loads(json_to_load)
-        return json_to_load
-    except json.JSONDecodeError as e:
-        if CFG.debug_mode:
-            print("json loads error", e)
-        error_message = str(e)
-        if error_message.startswith("Invalid \\escape"):
-            json_to_load = fix_invalid_escape(json_to_load, error_message)
-        if error_message.startswith(
-            "Expecting property name enclosed in double quotes"
-        ):
-            json_to_load = add_quotes_to_property_names(json_to_load)
-            try:
-                json.loads(json_to_load)
-                return json_to_load
-            except json.JSONDecodeError as e:
-                if CFG.debug_mode:
-                    print("json loads error - add quotes", e)
-                error_message = str(e)
-        if balanced_str := balance_braces(json_to_load):
-            return balanced_str
-    return json_to_load
 
 
 def fix_and_parse_json(
